+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 * Fix error during WordPress installation
 * Add `WP_STASH_BYPASS` environment variable
 * object-cache.php drop-in now merely requires the actual drop-in from the install folder. This makes it update-safe
-
+* Add a couple of type assertion for improved IDE support in object-cache.php
 #### 3.1.0
 * Implementation of `Debug\ActionLogger`
 * Added new Logger to `Pool`